% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_CRDscore.R
\name{cal_CRDscore}
\alias{cal_CRDscore}
\title{Calculate Circadian rhythm disruption score using scRNA or bulk transcriptomic data}
\usage{
cal_CRDscore(
  expr,
  circadians = NULL,
  study.type = NULL,
  num.rounds = 1000,
  n.bins = 50,
  seed = TRUE
)
}
\arguments{
\item{expr}{log2 transformed TPM matrix derived from scRNAseq, or normalized gene expression matrix from bulk samples; rows are genes, cols are cells or samples.}

\item{circadians}{The character vectors containing selected circadian features.}

\item{study.type}{A string value to indicate the study type for calculation. Allowed values contain c('scRNAseq', 'bulk_RNAseq').}

\item{num.rounds}{A integer value to indicate the permutation times of iteration; 1000 by default and 10000 will be better for reproducibility.}

\item{n.bins}{A integer value giving expression bins within genes, which were binned according to their average expression across cells or samples. Set to 50 by default.}

\item{seed}{A logical value to indicate if set random seed for reproducible results. TRUE by default.}
}
\value{
Continuous numerical variables with CRDscore, named by cells or samples.
}
\description{
Calculate Circadian rhythm disruption score using scRNA or bulk transcriptomic data
}
\examples{
\dontrun{
data("exprset")
data("circadians_selected")
CRDscore <- cal_CRDscore(expr = exprset, circadians = circadians_selected, study.type = "scRNAseq")
}
}
\author{
Leihe, Yixian Fan.
}
