% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadians_total.R
\docType{data}
\name{circadians_total}
\alias{circadians_total}
\title{Circadian-related genes}
\format{
A dataframe with 2230 rows (circadian-related genes) and 3 cols including "nums_circadian", "human_genesymbol", and "mice_genesymbol".
}
\source{
\url{http://circadb.org}
}
\usage{
circadians_total
}
\description{
A data frame storing 2230 circadian-related genes
}
\keyword{datasets}
