% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprset.R
\docType{data}
\name{exprset}
\alias{exprset}
\title{gene expression demo
A matrix frame storing gene expression data.}
\format{
A matrix with 16522 rows and 100 variables.
}
\usage{
exprset
}
\description{
gene expression demo
A matrix frame storing gene expression data.
}
\keyword{datasets}
